+++ Project1/Project1.vcxproj
@@ -28,12 +28,21 @@
     <ClInclude Include="Octree.h" />
     <ClInclude Include="Ray.h" />
     <ClInclude Include="Triangle.h" />
+    <ClInclude Include="Utility.h" />
     <ClInclude Include="Vec.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="Box.cpp" />
+    <ClCompile Include="BVH.cpp" />
+    <ClCompile Include="Camera.cpp" />
     <ClCompile Include="main.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">false</ExcludedFromBuild>
     </ClCompile>
+    <ClCompile Include="Material.cpp" />
+    <ClCompile Include="Mesh.cpp" />
+    <ClCompile Include="Octree.cpp" />
+    <ClCompile Include="Triangle.cpp" />
+    <ClCompile Include="Utility.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>15.0</VCProjectVersion>
